add scalajs support
scala.collection.concurrent.TrieMap is the only underlying
map type that fails to work on Scala.js (linking errors)
