+++ core/js/src/main/scala/io/chrisdavenport/mapref/MapRefCompanionPlatform.scala
@@ -1,0 +1,3 @@
+package io.chrisdavenport.mapref
+
+private[mapref] trait MapRefCompanionPlatform
