+++ project/plugins.sbt
@@ -4,7 +4,7 @@ addSbtPlugin("com.typesafe"              % "sbt-mima-plugin"          % "0.6.1")
 addSbtPlugin("org.scalameta"             % "sbt-mdoc"                 % "2.2.21")
 addSbtPlugin("com.47deg"                 % "sbt-microsites"           % "1.3.2")
 addSbtPlugin("com.typesafe.sbt"          % "sbt-ghpages"              % "0.6.3")
-addSbtPlugin("org.portable-scala"        % "sbt-scalajs-crossproject" % "0.6.1")
-addSbtPlugin("org.scala-js"              % "sbt-scalajs"              % "0.6.31")
+addSbtPlugin("org.portable-scala"        % "sbt-scalajs-crossproject" % "1.0.0")
+addSbtPlugin("org.scala-js"              % "sbt-scalajs"              % "1.5.1")
 addSbtPlugin("io.chrisdavenport"         % "sbt-mima-version-check"   % "0.1.2")
 addSbtPlugin("com.codecommit"            % "sbt-github-actions"       % "0.10.1")
